requirements.txt
opencv-contrib-python==4.0.0.21
pillow
h5py
scipy
numpy
tensorflow-gpu==2.5
scikit-learn
matplotlib
scikit-image
xlwt
onnx==1.18.0
onnxruntime==1.19.0
onnxsim==0.4.36
